\name{tiledbvcf-package}
\alias{tiledbvcf-package}
\alias{tiledbvcf}
\docType{package}
\title{\packageTitle{tiledbvcf}}
\description{\packageDescription{tiledbvcf}}
\section{Package Content}{\packageIndices{tiledbvcf}}
\author{\packageAuthor{tiledbvcf}}
\section{Maintainer}{\packageMaintainer{tiledbvcf}}
\keyword{package}
